requirements.txt
flask
flask-cors
requests
openai
python-dotenv
pymongo
requests
trafilatura
google-generativeai
scikit-learn
rapidfuzz
google-generativeai>=0.3.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
numpy>=1.24.0
pymongo>=4.0.0